Doc: Updating readme Build: Removing colors from the build Fix: Removing all references to colors to avoid collisions
