--- src/Components/Atoms/Colors/index.ts
@@ -1,3 +1,0 @@
-export { default as colors, COLORSENUM } from './colors';
-export { colorConverter } from './colorUtils';
-export type { IColor } from './colorUtils';
