+++ src/Components/Molecules/MultiSelectField/MultiSelectField.tsx
@@ -2,7 +2,6 @@ import React, { ReactElement, Ref } from 'react';
 import classnames from 'classnames';
 
 import { GenericField, IOption, MultiSelectInput } from '../../Atoms';
-import colors from '../../Atoms/Colors/colors';
 
 export interface IMultiSelectFieldProps {
   /** Custom colors settings */
@@ -140,11 +139,11 @@ export const MultiSelectField = (props: IMultiSelectFieldProps): ReactElement =>
 
 MultiSelectField.defaultProps = {
   colors: {
-    controlErrorColor: colors.error.rgb,
-    controlFocusColor: colors.primary.rgb,
+    controlErrorColor: 'rgb(255, 52, 24)',
+    controlFocusColor: 'rgb(38, 186, 212)',
     fontColor: 'rgb(0, 0, 0)',
-    optionFocusColor: colors.chalk.rgb,
-    optionSelectedColor: colors.primary.rgb,
+    optionFocusColor: 'rgb(228, 228, 228)',
+    optionSelectedColor: 'rgb(38, 186, 212)',
   },
   disabled: false,
   errorMessage: undefined,
