+++ src/assets/_global.scss
@@ -1,22 +1,22 @@
 @import './actionbar';
-@import './amount.scss';
-@import './badge.scss';
-@import './button.scss';
-@import './checkbox.scss';
-@import './datapicker.scss';
-@import './datatable.scss';
-@import './field.scss';
-@import './general.scss';
-@import './icon.scss';
-@import './layout.scss';
-@import './link.scss';
-@import './modal.scss';
-@import './navbar.scss';
-@import './popover.scss';
-@import './select.scss';
-@import './sidebar.scss';
-@import './switch.scss';
-@import './textarea.scss';
-@import './textinput.scss';
-@import './tooltip.scss';
-@import './typography.scss';
+@import './amount';
+@import './badge';
+@import './button';
+@import './checkbox';
+@import './datapicker';
+@import './datatable';
+@import './field';
+@import './general';
+@import './icon';
+@import './layout';
+@import './link';
+@import './modal';
+@import './navbar';
+@import './popover';
+@import './select';
+@import './sidebar';
+@import './switch';
+@import './textarea';
+@import './textinput';
+@import './tooltip';
+@import './typography';
